@@ -10,12 +10,12 @@
   "dependencies": {
     "core-js": "^3.6.5",
     "onecolor": "^3.1.0",
-    "vue": "^2.6.12",
-    "vue-class-component": "^7.2.6",
+    "vue": "^3.0.0-beta.1",
+    "vue-class-component": "^8.0.0-beta.3",
     "vue-gapi": "^0.3.1",
     "vue-gtag": "^1.9.1",
     "vue-property-decorator": "^9.0.0",
-    "vuex": "^3.5.1"
+    "vuex": "^4.0.0-alpha.1"
   },
   "devDependencies": {
     "@types/gapi.client": "^1.0.3",
@@ -27,6 +27,7 @@
     "@vue/cli-plugin-eslint": "^4.5.6",
     "@vue/cli-plugin-typescript": "^4.5.6",
     "@vue/cli-service": "^4.5.6",
+    "@vue/compiler-sfc": "^3.0.0-beta.1",
     "@vue/eslint-config-standard": "^5.1.2",
     "@vue/eslint-config-typescript": "^5.1.0",
     "eslint": "^7.9.0",
@@ -34,8 +35,7 @@
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.1",
-    "eslint-plugin-vue": "^6.2.2",
-    "typescript": "~4.0.3",
-    "vue-template-compiler": "^2.6.12"
+    "eslint-plugin-vue": "^7.0.0-alpha.0",
+    "typescript": "~4.0.3"
   }
 }
