@@ -6,18 +6,26 @@
 </template>
 
 <script lang="ts">
-import { Component, Prop, Vue } from 'vue-property-decorator'
-import { getRandomColor } from '@/utils'
+import { getRandomColor } from '@/utils';
+import { props } from "vue-class-component";
+import { PropType } from 'vue';
 import MediaItem = gapi.client.photoslibrary.MediaItem;
 
 const BACKGROUND_COLOR = '#202124'
 const BACKGROUND_COLOR_DELTA = 0.1
 
-@Component
-export default class AlbumMediaItemImage extends Vue {
-  @Prop(String) readonly descriptionTheme!: string;
-  @Prop(Object) readonly mediaItem!: MediaItem;
+const Props = props({
+  descriptionTheme: {
+    type: String,
+    required: true
+  },
+  mediaItem: {
+    type: Object as PropType<MediaItem>,
+    required: true
+  }
+});
 
+export default class AlbumMediaItemImage extends Props {
   backgroundColor: string = getRandomColor(BACKGROUND_COLOR, BACKGROUND_COLOR_DELTA);
 }
 </script>
